Tests/LibWeb/Text/expected/css/computed-values-group-sharing.txt

unstyled child animation at default: false
unstyled child grid at default: false
unstyled child svgReset at default: true
unstyled child mask at default: false
unstyled child anchor at default: true
unstyled child content at default: true
unstyled child transform at default: false
unstyled child effects at default: true
unstyled child inheritedTable at default: true
unstyled child inheritedList at default: false
unstyled child inheritedSVG at default: false
color-styled child inheritedText at default: false
color-styled child inheritedText shared with parent: false
color-styled child grid still at default: false
grid-styled child grid at default: false
